+++ ExperimentalEnemyInteractions/CHANGELOG.md
@@ -1,5 +1,11 @@
 <h1>Natural selection</h1>
 
+0.2.0 <br>
+	- **Added networking** <br>
+	- - Fixes desync and network related issues, mainly enemy sounds not playing on non-host clients<br>
+	- Fixed sandworm eating non-host players inside the ship with the toggle on<br>
+	- Fixed giants not using the extinguish chance from config<br>
+	- Added __blacklist__ to config<br>
 0.1.5 <br>
 	- Fixed collisions returning NullReferenceException errors<br>
 	- Added config for sandworms to not eat players inside leaving ship<br>
@@ -27,6 +33,9 @@
 
 <h1>Experimental Enemy Interactions</h1>
 
+0.3.0 - <br>
+	- See 0.2.0
+	- NaturalSelectionLib 0.6.0 included
 0.2.6 - <br>
 	- Finally discovered a fix for bees <br>
 	- Hygrodiges move towards and consume corpses<br>